requirements.txt
discord.py == 2.6.4
python-dotenv == 1.2.1
PyYAML == 6.0.2
